// frontend/src/api-client/models/Animal.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VET API
 * CRM pour vétérinaires
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Client } from './Client';
import {
    ClientFromJSON,
    ClientFromJSONTyped,
    ClientToJSON,
} from './Client';

/**
 * 
 * @export
 * @interface Animal
 */
export interface Animal {
    /**
     * 
     * @type {number}
     * @memberof Animal
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Animal
     */
    firstName: string;
    /**
     * 
     * @type {number}
     * @memberof Animal
     */
    age: number;
    /**
     * 
     * @type {number}
     * @memberof Animal
     */
    height: number;
    /**
     * 
     * @type {number}
     * @memberof Animal
     */
    weight: number;
    /**
     * 
     * @type {string}
     * @memberof Animal
     */
    species: AnimalSpeciesEnum;
    /**
     * 
     * @type {Client}
     * @memberof Animal
     */
    client: Client;
    /**
     * 
     * @type {number}
     * @memberof Animal
     */
    clientId: number;
    /**
     * 
     * @type {Date}
     * @memberof Animal
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Animal
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const AnimalSpeciesEnum = {
    Chien: 'chien',
    Chat: 'chat',
    Hamster: 'hamster',
    Tortue: 'tortue'
} as const;
export type AnimalSpeciesEnum = typeof AnimalSpeciesEnum[keyof typeof AnimalSpeciesEnum];


/**
 * Check if a given object implements the Animal interface.
 */
export function instanceOfAnimal(value: object): boolean {
    if (!('id' in value)) return false;
    if (!('firstName' in value)) return false;
    if (!('age' in value)) return false;
    if (!('height' in value)) return false;
    if (!('weight' in value)) return false;
    if (!('species' in value)) return false;
    if (!('client' in value)) return false;
    if (!('clientId' in value)) return false;
    if (!('createdAt' in value)) return false;
    if (!('updatedAt' in value)) return false;
    return true;
}

export function AnimalFromJSON(json: any): Animal {
    return AnimalFromJSONTyped(json, false);
}

export function AnimalFromJSONTyped(json: any, ignoreDiscriminator: boolean): Animal {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'firstName': json['firstName'],
        'age': json['age'],
        'height': json['height'],
        'weight': json['weight'],
        'species': json['species'],
        'client': ClientFromJSON(json['client']),
        'clientId': json['clientId'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function AnimalToJSON(value?: Animal | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'firstName': value['firstName'],
        'age': value['age'],
        'height': value['height'],
        'weight': value['weight'],
        'species': value['species'],
        'client': ClientToJSON(value['client']),
        'clientId': value['clientId'],
        'createdAt': ((value['createdAt']).toISOString()),
        'updatedAt': ((value['updatedAt']).toISOString()),
    };
}